+++ resources/views/components/form-label.blade.php
@@ -1,7 +1,7 @@
 @if ($label)
     <div class="label_wrap">
         <label {!! $attributes->merge() !!}>
-            {{ $label }}@if ($required)<span class="required">*</span>@endif
+            {!! $label !!}@if ($required)<span class="required">*</span>@endif
         </label>
     </div>
 @endif