+++ resources/views/components/form-select.blade.php
@@ -1,7 +1,9 @@
-<div class="field_wrap {{ implode(' ', $type) }}">
+@if ($field_wrap)
+    <x-form-wrap field-identifier="{{ implode(' ', $type) }}">
+@endif
 
     <x-form-label label="{{ $label }}" :required="$required" for="{{ $id() }}" class="{{ $hasError() ? 'error' : '' }}" />
-    
+
     <div class="input_wrap {{ $hasError() ? 'error' : '' }}">
         <select
             id="{{ $id() }}"
@@ -20,11 +22,14 @@
             @endforeach
         </select>
     </div>
-    
+
     @if ($showError())
         <x-form-error name="{{ $name }}" />
     @endif
-</div>
+
+@if ($field_wrap)
+    </x-form-wrap>
+@endif
 
 @push('scripts')
 <script type="text/javascript">