+++ resources/views/components/form-wrap.blade.php
@@ -1,0 +1,5 @@
+<div class="field_wrap {{ $field_identifier }}">
+
+    {!! $slot !!}
+
+</div>