+++ src/View/Components/FormError.php
@@ -8,13 +8,12 @@ use Str;
 class FormError extends Component
 {
     public $name;
-    
+
     public function __construct($name)
     {
-        
         $this->name = str_replace(['[', ']'], ['.', ''], Str::before($name, '[]'));
     }
-    
+
     public function render()
     {
         return view(config('spin-forms.components.form-error.view'));