+++ src/View/Components/FormInput.php
@@ -4,20 +4,21 @@ namespace Spindogs\LaravelSpinForms\View\Components;
 
 class FormInput extends FormElement
 {
+    public $field_wrap;
     public $label;
     public $name;
     public $required;
     public $show_error;
     public $type;
-    public $value;
 
     public function __construct(
         string $name,
         string $id = null,
-        string $label = '',
+        string $label = null,
         string $type = 'text',
-        $default = null,
+        $value = null,
         bool $required = false,
+        bool $fieldWrap = true,
         bool $showError = true
     ) {
         $this->name         = $name;
@@ -26,7 +27,8 @@ class FormInput extends FormElement
         $this->type         = $type;
         $this->required     = $required;
         $this->show_error   = $showError;
+        $this->field_wrap   = $fieldWrap;
 
-        $this->value        = old($name, $default);
+        $this->setValue($name, $value);
     }
 }