+++ changelog.md
@@ -2,6 +2,16 @@
 
 All notable changes to `LaravelSpinForms` will be documented in this file.
 
+## Version 2.0.0
+
+### New Version
+- Added in Form Wrap, Form Group.
+- Field wrap element is decoupled html structure allowing the separation of label and element where necessary.
+- Radio and Checkbox have been rewritten. The no longer take array inputs, but are now designed to sit inside a Form Group. This is a breaking change from V1 to V2
+- Changed "default" to "value" when calling elements to make more semantic sense.
+- Refactored code
+- Minor bug fixes
+
 ## Version 1.0.2
 
 ### Added